==> 2_GenerateTargetSummaries.ipynb <==
# Generate Target Summaries

Use OpenAI GPT-4o to generate the target summaries our models will be trained on.

# %pip install openai
# %pip install --upgrade openai

## Read the kaggle dataset

import pandas as pd

dataset_dir = 'datasets'

# read the dataset
df = pd.read_csv(f'{dataset_dir}/podcast.csv')
df.head()

# we only want the text column
df = df[['text']]
df = df.dropna()

## Generate the target summaries by calling openai

from openai import OpenAI

api_key = ""
client = OpenAI(
    api_key=api_key,
)

from SharedUtils import trim_to_max_length

# trim the long length summary to a shorter length
df['text_short'] = df['text'].apply(lambda x: trim_to_max_length(x))
df.head()

df['text_short'][126]

prompt = "Write a short summary, limited to 20 words, for the following podcast transcript: "

def generate_summary(text, index):
    # gpt3 has a maximum context length is 16385 tokens.
    # I guesstimated the characters that allow for 16385 tokens
    # max_text_length = 65000
    # if len(text) > max_text_length:
    #     text = text[:max_text_length]

    try:
        prompt_to_send = prompt + "\"" + text + "\""
        chat_completion = client.chat.completions.create(
            messages=[
                {
                    "role": "user",
                    "content": prompt_to_send,
                }
            ],
            # model="gpt-3.5-turbo",
            model="gpt-4o"
        )
        response_summary = chat_completion.choices[0].message.content
        # remove "Summary:" from the response
        response_summary = response_summary.replace("Summary:", "")
        return response_summary
    except Exception as e:
        print(f"Error generating summary for text: {e}, Index {index}")
        print(f"Text Length: {len(text)}")
        return None

# generate row 1 summary
print(generate_summary(df['text_short'][126], 126))

import time
start_time = time.time()

df['summary'] = df['text_short'].apply(lambda x: generate_summary(x, df['text_short'].index[df['text_short'] == x][0]))

end_time = time.time()
elapsed_time = end_time - start_time
print(f"Time taken to generate {df.shape[0]} summaries: {elapsed_time:.2f} seconds")

# get elapsed time in minutes
elapsed_time_minutes = elapsed_time / 60
print(f"Time taken to generate {df.shape[0]} summaries: {elapsed_time_minutes:.2f} minutes")

# do a second pass to generate a second round of summaries
# we can use these to gather metrics (rouge, bleu, etc) on 4o summaries against itself
df['summary2'] = df['text_short'].apply(lambda x: generate_summary(x, df['text_short'].index[df['text_short'] == x][0]))

df.head()

# save the dataset
df.to_csv(f'{dataset_dir}/podcast_with_summary.csv', index=False)

# also save a train and test dataset
from sklearn.model_selection import train_test_split

train, test = train_test_split(df, test_size=0.2, random_state=42)

train.to_csv(f'{dataset_dir}/podcast_with_summary_train.csv', index=False)
test.to_csv(f'{dataset_dir}/podcast_with_summary_test.csv', index=False)

## Generate baseline metrics that can be compared against

Calculate the rouge and bleu for 4o summaries vs 4o summaries

from SharedUtils import evaluate_and_save_metrics

reference_summaries = df['summary'].tolist()
generated_summaries = df['summary2'].tolist()

rouge_results, bleu_results = evaluate_and_save_metrics(
    "openai-gpt-4o",
    "baseline",
    "gpt_4o",
    reference_summaries,
    generated_summaries,
    elapsed_time
)

print(rouge_results)
print(bleu_results)
