==> 4_Seq2Seq.ipynb <==
# T5 Fine Tune

Fine tune the Google T5 with the summary dataset and evaluate the model

import pandas as pd

dataset_dir = 'datasets'

# read the dataset
df = pd.read_csv(f'{dataset_dir}/podcast_with_summary_train.csv')
df.head()

# find the summary with the longest length
df['summary_length'] = df['summary'].apply(lambda x: len(x))
print(df['summary_length'].max())

# find the text with the longest length
df['text_length'] = df['text_short'].apply(lambda x: len(x))
print(df['text_length'].max())

from datasets import Dataset
from transformers import AutoTokenizer, AutoModelForSeq2SeqLM, Trainer, TrainingArguments

dataset = Dataset.from_pandas(df)

model_name = "t5-small"
# 60.5M params
tokenizer = AutoTokenizer.from_pretrained(model_name)

def preprocess_data(examples):
    # Tokenize inputs
    model_inputs = tokenizer(
        examples["text_short"], max_length=1024, truncation=True, padding="max_length"
    )
    # Tokenize targets
    labels = tokenizer(
        examples["summary"], max_length=200, truncation=True, padding="max_length"
    )
    
    # Replace padding token id's of the labels by -100
    labels["input_ids"] = [
        [(l if l != tokenizer.pad_token_id else -100) for l in label]
        for label in labels["input_ids"]
    ]
    model_inputs["labels"] = labels["input_ids"]
    return model_inputs


# Apply the preprocessing
tokenized_dataset = dataset.map(preprocess_data, batched=True)

model = AutoModelForSeq2SeqLM.from_pretrained(model_name)

training_args = TrainingArguments(
    output_dir="./tmp/t5_output",
    evaluation_strategy="steps",
    eval_steps=500,
    logging_steps=100,
    learning_rate=2e-5,
    per_device_train_batch_size=4,  
    per_device_eval_batch_size=4,
    weight_decay=0.01,
    save_total_limit=2,
    num_train_epochs=100,  
    logging_dir='./logs', 
    report_to="none", 
)

from transformers import DataCollatorForSeq2Seq, Trainer
import numpy as np

data_collator = DataCollatorForSeq2Seq(
    tokenizer, model=model, label_pad_token_id=-100
)

trainer = Trainer(
    model=model,
    args=training_args,
    train_dataset=tokenized_dataset,
    eval_dataset=tokenized_dataset,
    tokenizer=tokenizer,
    data_collator=data_collator,
)

trainer.train()

def run_inference(text):
    inputs = tokenizer(
        text,
        return_tensors="pt",
        truncation=True,
        padding="max_length",
        max_length=1024
    )
    device = model.device
    inputs = {k: v.to(device) for k, v in inputs.items()}  # Move inputs to GPU

    summary_ids = model.generate(
        inputs["input_ids"],
        max_length=200, 
        num_beams=4,
        early_stopping=True
    )
    summary = tokenizer.decode(summary_ids[0], skip_special_tokens=True)
    return summary

test_df = pd.read_csv(f'{dataset_dir}/podcast_with_summary_test.csv')

test_text = test_df.text_short[0]
print("Original:", test_text)

summary = run_inference(test_text)

print("Predicted:", summary)
print("Actual:", test_df.summary[0])

import time
from SharedUtils import evaluate_and_save_metrics

# iterate over all rows in test_df
def evaluate_df(df, name):
    reference_summaries = []
    predicted_summaries = []
    total_time = 0
    for i, row in df.iterrows():
        test_text = row.text_short
        
        start_time = time.time()
        summary = run_inference(test_text)
        end_time = time.time()
        elapsed_time = end_time - start_time
        total_time += elapsed_time
        
        reference_summaries.append(row.summary)
        predicted_summaries.append(summary)
    
    model_name = "t5-small"

    rouge_results, results_bleu = evaluate_and_save_metrics(
        model_name,
        name,
        "finetuned",
        reference_summaries,
        predicted_summaries,
        total_time
    )

    results_df = pd.DataFrame({
        'summary': reference_summaries,
        'summary_tuned': predicted_summaries
    })
    results_df.to_csv(f"./results/{model_name}/{name}/summaries.csv")

    print(rouge_results)
    print(results_bleu)
    print("Total time (seconds):", total_time)
    print("Total time (minutes): ", total_time / 60)
    

evaluate_df(test_df, "test_dataset")

train_df = pd.read_csv(f'{dataset_dir}/podcast_with_summary_train.csv')
evaluate_df(train_df, 'train_dataset')

whole_df = pd.read_csv(f'{dataset_dir}/podcast_with_summary.csv')
evaluate_df(whole_df, 'whole_dataset')
