==> 5_SummaryUsingSmallGPT.ipynb <==
# Fine-tune a smaller GPT model

Train a smaller GPT model to produce summaries like the gpt-4o model.

## Notebook controls

# notebook controls
remove_old_directories = True
train_model = True
evaluate_model = True

## Remove artifacts from previous run

import shutil

# remove all directories from training and testing
if (remove_old_directories):
    shutil.rmtree('./results/gpt', ignore_errors=True)
    shutil.rmtree('./tuned_model', ignore_errors=True)
    shutil.rmtree('./tuned_tokenizer', ignore_errors=True)
    shutil.rmtree('./tmp', ignore_errors=True)

## Select the base model

# model_name = "facebook/opt-350m"
# model_name = "meta-llama/Llama-2-7b-hf"
# model_name = "openai-community/gpt2"
model_name = "EleutherAI/gpt-neo-125m"


## Train/fine-tune the gpt model to produce summaries 

from SharedUtils import trim_to_max_length

def hashtag_format(text, summary):
    prompt = f"### Question: Summarize the following which is surrounded by quotes \"{text}\"\n### Answer:"
    if (len(summary) > 0):
        prompt += f" {summary}"
    return prompt

def format_text(text, summary):
    limited_text = trim_to_max_length(text)
    return hashtag_format(limited_text, summary)

response_template = "### Answer:"
    

from transformers import AutoModelForCausalLM, AutoTokenizer, TrainingArguments
from datasets import load_dataset
from trl import SFTTrainer, DataCollatorForCompletionOnlyLM

if (train_model):
    dataset = load_dataset("csv", data_files="./datasets/podcast_with_summary_train.csv", split="train")

    model = AutoModelForCausalLM.from_pretrained(model_name)
    tokenizer = AutoTokenizer.from_pretrained(model_name)
    tokenizer.pad_token = tokenizer.eos_token

    def formatting_prompts_func(example):
        output_texts = []
        for i in range(len(example['text_short'])):
            txt = example['text_short'][i]
            text = format_text(txt, example['summary'][i])
            output_texts.append(text)
        
        return output_texts

    collator = DataCollatorForCompletionOnlyLM(response_template, tokenizer=tokenizer)

    training_args = TrainingArguments(
        output_dir="./tmp",
        evaluation_strategy="steps",
        eval_steps=500,
        logging_dir='./logs',
        logging_steps=100,
        save_steps=500,
        save_total_limit=2,
        per_device_train_batch_size=4,
        per_device_eval_batch_size=4,
        num_train_epochs=3,
        weight_decay=0.01,
        learning_rate=5e-5,
        gradient_accumulation_steps=8
        )

    trainer = SFTTrainer(
        model,
        train_dataset=dataset,
        eval_dataset=dataset,
        args=training_args,
        formatting_func=formatting_prompts_func,
        data_collator=collator,
    )

    trainer.train()

    trainer.save_model("./tuned_model")
    tokenizer.save_pretrained("./tuned_tokenizer")

## Test inference using the new fine tuned model

import pandas as pd

df = pd.read_csv("./datasets/podcast_with_summary_test.csv")
first_text = df['text'][0]
expected_summary = df['summary'][0]

prompt = format_text(first_text, "")

print(len(prompt))
print(prompt)
print(expected_summary)

from SharedUtils import trim_to_last_punctuation

def parse_summary_from_response(response):
    # find the start token
    response_template_trimmed = response_template.strip()
    start_token = response.find(response_template_trimmed)
    if start_token == -1:
        return ""
    # trim from end of the start token to the end of the response
    result = response[start_token + len(response_template_trimmed):]
    result = trim_to_last_punctuation(result)    
    return result

max_new_tokens = 48

from transformers import AutoModelForCausalLM, AutoTokenizer, TrainingArguments
from transformers import pipeline
import time

def create_generator(model_name, tokenizer_name):
    model = AutoModelForCausalLM.from_pretrained(model_name)
    tokenizer = AutoTokenizer.from_pretrained(tokenizer_name)
    tokenizer.pad_token = tokenizer.eos_token
    return pipeline("text-generation", model=model, tokenizer=tokenizer, device="cuda", truncation=True)

def run_inference(generator, prompt, max_new_tokens, log_results):
    elapsed_time = 0
    start_time = time.time()
    generated_text = generator(prompt, max_new_tokens=max_new_tokens, num_return_sequences=1)
    end_time = time.time()
    elapsed_time = end_time - start_time
    if log_results:
        print(generated_text[0]["generated_text"])
        print("\n")
    return parse_summary_from_response(generated_text[0]["generated_text"]), elapsed_time

generator_finetuned = create_generator("./tuned_model", "./tuned_tokenizer")
summary, elapsed = run_inference(generator_finetuned, prompt, max_new_tokens, True)

print("Summary: " + summary)

generator_original = create_generator(model_name, model_name)
summary, elapsed = run_inference(generator_original, prompt, max_new_tokens, True)

print("Summary: " + summary)

## Evaluate the results
Compare the fine tuned model against the base model

import os
os.makedirs("./results", exist_ok=True)
# remove dir and all subdirs
shutil.rmtree("./results/gpt", ignore_errors=True)
os.makedirs("./results/gpt", exist_ok=True)

from SharedUtils import evaluate_and_save_metrics

def evaluate_df(df, generator, generator_tuned, max_new_tokens, name):
    if (not evaluate_model):
        return

    os.makedirs(f"./results/gpt/{name}", exist_ok=True)

    total_time_orig = 0
    total_time_tuned = 0
    summaries_orig = []
    summaries_tuned = []
    reference_summaries = []
    
    for i in range(len(df)):
        text = df['text_short'][i]
        expected_summary = df['summary'][i]
        prompt = format_text(text, "")

        summary, orig_time = run_inference(generator, prompt, max_new_tokens, False)
        total_time_orig += orig_time
        summaries_orig.append(summary)
        
        summary, tuned_time = run_inference(generator_tuned, prompt, max_new_tokens, False)
        total_time_tuned += tuned_time
        summaries_tuned.append(summary)
        
        reference_summaries.append(expected_summary)

    print("Original Model")
    rouge_results, bleu_results = evaluate_and_save_metrics(
        "gpt",
        name,
        "gpt_basemodel",
        reference_summaries,
        summaries_orig,
        total_time_orig
    )
    print(rouge_results)
    print(bleu_results)
    print(f"Total time (seconds): {total_time_orig}")
    print(f"Total time (minutes): {total_time_orig / 60}")

    print("Tuned Model")
    rouge_results, bleu_results = evaluate_and_save_metrics(
        "gpt",
        name,
        "gpt_tunedmodel",
        reference_summaries,
        summaries_tuned,
        total_time_tuned
    )
    print(rouge_results)
    print(bleu_results)
    print(f"Total time (seconds): {total_time_tuned}")
    print(f"Total time (minutes): {total_time_tuned / 60}")

    results_df = pd.DataFrame({
        'summary': reference_summaries,
        'summary_orig': summaries_orig,
        'summary_tuned': summaries_tuned
    })
    results_df.to_csv(f"./results/gpt/{name}/summaries.csv")

df = pd.read_csv("./datasets/podcast_with_summary_test.csv")
results_df = evaluate_df(df, generator_original, generator_finetuned, max_new_tokens, "test_dataset")

df = pd.read_csv("./datasets/podcast_with_summary_train.csv")
results_df = evaluate_df(df, generator_original, generator_finetuned, max_new_tokens, "train_dataset")

# read the entire dataset for the final evaluation
df = pd.read_csv("./datasets/podcast_with_summary.csv")
results_df = evaluate_df(df, generator_original, generator_finetuned, max_new_tokens, "whole_dataset")
