==> 6_MetricsAndResults.ipynb <==
# Metrics, Results, and Conclusions

Analyze all metrics and determine the best model

## Load the metrics that were output for all models

import json 

def read_time_file(file_path):
    with open(file_path, "r") as f:
        content = f.read()
    return float(content)

def read_results_files(file_path):    
    with open(file_path, "r") as f:
        results = json.load(f)    
    return results

import pandas as pd

# read lstm results
lstm_time_seconds = read_time_file("./results/pytorch/whole_dataset/100_epochs_time.txt")
lstm_bleu_results = read_results_files("./results/pytorch/whole_dataset/100_epochs_bleu_results.json")
lstm_rouge_results = read_results_files("./results/pytorch/whole_dataset/100_epochs_rouge_results.json")
lstm_summaries = pd.read_csv("./results/pytorch/whole_dataset/100_epochs_summaries.csv")

# read t5 results
t5_time_seconds = read_time_file("./results/t5-small/whole_dataset/gpt_finetuned_time.txt")
t5_bleu_results = read_results_files("./results/t5-small/whole_dataset/gpt_finetuned_bleu_results.json")
t5_rouge_results = read_results_files("./results/t5-small/whole_dataset/gpt_finetuned_rouge_results.json")
t5_summaries = pd.read_csv("./results/t5-small/whole_dataset/summaries.csv")

# read gpt results
gpt_time_seconds = read_time_file("./results/gpt/whole_dataset/gpt_tunedmodel_time.txt")
gpt_bleu_results = read_results_files("./results/gpt/whole_dataset/gpt_tunedmodel_bleu_results.json")
gpt_rouge_results = read_results_files("./results/gpt/whole_dataset/gpt_tunedmodel_rouge_results.json")
gpt_summaries = pd.read_csv("./results/gpt/whole_dataset/summaries.csv")

# read the target 4o results
gpt_4o_time_seconds = read_time_file("./results/openai-gpt-4o/baseline/gpt_4o_time.txt")
gpt_4o_bleu_results = read_results_files("./results/openai-gpt-4o/baseline/gpt_4o_bleu_results.json")
gpt_4o_rouge_results = read_results_files("./results/openai-gpt-4o/baseline/gpt_4o_rouge_results.json")
gpt_4o_summaries = pd.read_csv("./datasets/podcast_with_summary.csv")

def print_results(time, bleu, rouge, summaries, column_name):
    print("Time (secoonds): ", time)
    print("Time (minutes): ", time / 60)
    print("\n")
    print("BLEU: ", bleu)
    print("\n")
    print("ROUGE: ", rouge)
    print("\n")
    print("Example Summaries: \n")
    print(summaries[column_name][0], "\n")
    print(summaries[column_name][1], "\n")
    print(summaries[column_name][2], "\n")
    print(summaries[column_name][3], "\n")
    print(summaries[column_name][4], "\n")

### Baseline Target from LLM

# print 4o results
print("GPT 4o results")
print_results(gpt_4o_time_seconds, gpt_4o_bleu_results, gpt_4o_rouge_results, gpt_4o_summaries, "summary2")

### LSTM Results

# print lstm results
print("LSTM results:")
print_results(lstm_time_seconds, lstm_bleu_results, lstm_rouge_results, lstm_summaries, "summary_tuned")

#### Human Evaluation of Summaries

The LSTM summaries appear to be nonsensical, containing a random assortment of words and phrases without any coherent structure, making them ineffective for conveying the summary for the input text.

### T5 Results

# t5 results
print("Seq2Seq (T5) results:")
print_results(t5_time_seconds, t5_bleu_results, t5_rouge_results, t5_summaries, "summary_tuned")

#### Human Evaluation of Summaries

Overall, this set of summaries is a significant improvement over the previous batch. Three of the summaries are clear, concise, and informative, effectively capturing the summary of the text. However, two summaries lack clarity and context. One is overly philosophical and vague, while the other is an unframed question, making them ineffective.

### GPT Results

# print gpt results
print("GPT results:")
print_results(gpt_time_seconds, gpt_bleu_results, gpt_rouge_results, gpt_summaries, "summary_tuned")

### Human Evaluation

This batch of summaries shows some improvement in clarity and detail, with the first and fifth summaries providing reasonable focus on the guests and their contributions. However, repetition and vagueness detract from the overall quality, as seen in the redundant phrasing of Koch’s summary and the lack of context in the neural networks and life’s meaning summaries.

# Conclusion

The generated summaries still fall short of the quality achieved by the LLM, but with a larger and more diverse training set, their performance could potentially improve. Among the models tested, T5 demonstrated the most consistent output, while the GPT model produced the highest-quality summaries, albeit with some repetition. These shortcomings may be mitigated through additional data cleaning and fine-tuning steps.

From a performance perspective, smaller models offer significant advantages in terms of speed and cost, producing text much faster and at a fraction of the expense of an LLM. For cost-sensitive applications, using a smaller model fine-tuned on an LLM-derived dataset presents a compelling and practical alternative.

# generate some metrics on BLEU, ROUGE, and time
# percent worse or better than GPT 4o
def percent_worse_than_gpt_4o(time, gpt_4o_time):
    return abs((time - gpt_4o_time) / gpt_4o_time * 100)

def percent_better_than_gpt_4o(time, gpt_4o_time):
    return abs((gpt_4o_time - time) / gpt_4o_time * 100)

def percent_worse_than_gpt_4o_bleu(bleu, gpt_4o_bleu):
    return abs((bleu['bleu'] - gpt_4o_bleu['bleu']) / gpt_4o_bleu['bleu'] * 100)

def percent_worse_than_gpt_4o_rouge(rouge, gpt_4o_rouge):
    return abs((rouge['rouge1'] - gpt_4o_rouge['rouge1']) / gpt_4o_rouge['rouge1'] * 100)

# now calculate the metrics
print("Metrics:")
print("LSTM:")
print("Time percent worse than GPT 4o: ", round(percent_worse_than_gpt_4o(lstm_time_seconds, gpt_4o_time_seconds), 2), "%")
print("BLEU percent worse than GPT 4o: ", round(percent_worse_than_gpt_4o_bleu(lstm_bleu_results, gpt_4o_bleu_results), 2), "%")
print("ROUGE percent worse than GPT 4o: ", round(percent_worse_than_gpt_4o_rouge(lstm_rouge_results, gpt_4o_rouge_results), 2), "%")
print("\n")

print("Seq2Seq (T5):")
print("Time percent better than GPT 4o: ", round(percent_better_than_gpt_4o(t5_time_seconds, gpt_4o_time_seconds), 2), "%")
print("BLEU percent worse than GPT 4o: ", round(percent_worse_than_gpt_4o_bleu(t5_bleu_results, gpt_4o_bleu_results), 2), "%")
print("ROUGE percent worse than GPT 4o: ", round(percent_worse_than_gpt_4o_rouge(t5_rouge_results, gpt_4o_rouge_results), 2), "%")
print("\n")

print("GPT:")
print("Time percent better than GPT 4o: ", round(percent_better_than_gpt_4o(gpt_time_seconds, gpt_4o_time_seconds), 2), "%")
print("BLEU percent worse than GPT 4o: ", round(percent_worse_than_gpt_4o_bleu(gpt_bleu_results, gpt_4o_bleu_results), 2), "%")
print("ROUGE percent worse than GPT 4o: ", round(percent_worse_than_gpt_4o_rouge(gpt_rouge_results, gpt_4o_rouge_results), 2), "%")
print("\n")
